updates repository ID type

+++ src/main/kotlin/com/sample/repository/TodoRepository.kt
@@ -3,6 +3,6 @@ package com.sample.repository
 import com.sample.model.Todo
 import org.springframework.data.repository.CrudRepository
 
-interface TodoRepository : CrudRepository<Todo, Long> {
+interface TodoRepository : CrudRepository<Todo, Int> {
 
 }
